Mode model-assembler and model-verifier to ...assemble and ...verify packages.
