+++ model/model-assembler/src/main/java/io/spine/model/assemble/SpineAnnotationProcessor.java
@@ -18,7 +18,7 @@
  * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-package io.spine.model;
+package io.spine.model.assemble;
 
 import com.google.common.base.Optional;
