+++ model/model-assembler/src/main/java/io/spine/model/assemble/package-info.java
@@ -18,30 +18,10 @@
  * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-package io.spine.model;
-
-import com.google.protobuf.Any;
-import com.google.protobuf.UInt32Value;
-import io.spine.server.command.Assign;
-import io.spine.server.command.CommandHandler;
-import io.spine.server.event.EventBus;
-import io.spine.validate.AnyVBuilder;
-
-import java.util.List;
-
-import static java.util.Collections.singletonList;
-
 /**
- * @author Dmytro Dashenkov
+ * This package provides tools for assembling the Spine model at compile time.
  */
-public class TestCommandHandler extends CommandHandler {
+@ParametersAreNonnullByDefault
+package io.spine.model.assemble;
 
-    protected TestCommandHandler(EventBus eventBus) {
-        super(eventBus);
-    }
-
-    @Assign
-    public List<UInt32Value> handle(UInt32Value command) {
-        return singletonList(command);
-    }
-}
+import javax.annotation.ParametersAreNonnullByDefault;
