+++ model/model-assembler/src/test/java/io/spine/model/assemble/SpineAnnotationProcessorShould.java
@@ -18,7 +18,7 @@
  * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-package io.spine.model;
+package io.spine.model.assemble;
 
 import org.junit.Before;
 import org.junit.Test;
