+++ model/model-verifier/src/main/resources/META-INF/gradle-plugins/io.spine.tools.spine-model-verifier.properties
@@ -19,4 +19,4 @@
 # OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 #
 
-implementation-class=io.spine.model.ModelVerifierPlugin
+implementation-class=io.spine.model.verify.ModelVerifierPlugin
