+++ model/model-verifier/src/test/resources/model-verifier-test/src/main/java/io/spine/model/verify/TestAggregate.java
@@ -18,39 +18,26 @@
  * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-package io.spine.model;
+package io.spine.model.verify;
 
-import com.google.protobuf.Any;
-import com.google.protobuf.UInt64Value;
 import io.spine.server.aggregate.Aggregate;
 import io.spine.server.command.Assign;
 import io.spine.validate.AnyVBuilder;
 
+import com.google.protobuf.Any;
+import com.google.protobuf.Message;
+
+import java.util.Collections;
 import java.util.List;
 
-import static java.util.Collections.singletonList;
+class TestAggregate extends Aggregate<String, Any, AnyVBuilder> {
 
-/**
- * @author Dmytro Dashenkov
- */
-public class DuplicateAggregate extends Aggregate<String, Any, AnyVBuilder> {
-
-    protected DuplicateAggregate(String id) {
+    public TestAggregate(String id) {
         super(id);
     }
 
     @Assign
-    public List<UInt64Value> handle(UInt64Value command) {
-        return singletonList(command);
-    }
-
-    @Assign
-    public List<Any> on(Any command) {
-        return singletonList(command);
-    }
-
-    @Assign
-    public void oneMore(Any cmd) {
-        // NoOp for test
+    List<Message> handle(Any command) {
+        return Collections.emptyList();
     }
 }
